Implemented toggle_pytest_ini + creating pytest.ini
This allows you to toggle on/off between whether you want the download function to also create a pytest.ini file in the program's directory (Which I implemented in this commit too). This removes those annoying warnings of a custom mark whenever you want to test a program.

Alternatively, you can switch this in sublime-settings yourself.