Fixed Open + Flags Comments
I noticed that Open didn't seem to want to work, and realized that it was prolly based off the old version of exercism's open command. Thus, I went in, fixed the match check, and rewrote the code to work with the new command and contain the proper failchecks.

I also fixed the comments that tell you about the custom flags for pytest to show the actual commands.